Rename neighbour header to mesh_neighbours.h
