+++ HEN_HOUSE/egs++/geometry/egs_mesh/tests/Makefile
@@ -2,5 +2,5 @@ CXX      = g++
 CXXFLAGS = -Wall -Wextra -O2 -std=c++11
 
 all: neighbour-test
-neighbour-test: neighbour-test.cpp ../neighbour.h
+neighbour-test: neighbour-test.cpp ../mesh_neighbours.h
 	$(CXX) $(CXXFLAGS) neighbour-test.cpp -I../ -o neighbour-test
